requirements.txt
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
scikit-learn>=1.5.0
tqdm>=4.66.1
python-dotenv>=1.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0

